- removed enum STATE_TYPE because we solved this better - removed tape_action.h, tape_action.c and moved this functionality to   tape_actions completely - replaced data member of tape_actions with three arrays:     vec_read - array of uintptr_t pointers (read vector)     vec_write - array of uintptr_t pointers (write vector)     vec_dirs - array of SHIFT_DIR enums (direction vector - removed priv member from tape - added safety checks in state.c tape_cell_list.c, tape_cell.c,   tape_actions.c, tm.c, edge_list.c, state_list.c (almost every .c-file)   tape.c, tapes.c - removed functions which used edge_default like tapes_apply_actions_default... - removed alphabet

+++ include/tape_cell.h
@@ -1,6 +1,8 @@
 #ifndef TAPE_CELL_H_
 #define TAPE_CELL_H_
 
+#include <stdint.h>
+
 /**
  * \struct tape_cell
  * \brief Represents the "Node/Entry Class" of #tape_cell_list
@@ -13,12 +15,12 @@
  * Prev-Pointer
  */
 struct tape_cell {
-	unsigned int token;
+	uintptr_t token;
 	struct tape_cell *next;
 	struct tape_cell *prev;
 };
 
-struct tape_cell *tape_cell_new(unsigned int token);
+struct tape_cell *tape_cell_new(uintptr_t token);
 struct tape_cell *tape_cell_copy(struct tape_cell *this);
 void tape_cell_free(struct tape_cell *this);
 void tape_cell_print(struct tape_cell *this);
